Decompose the extensions file into separate files for diagnostic scopes
